Relocated ConnectionBridge and PacketEncoder/Decoder
