+++ common/src/main/java/org/adde0109/pcf/forwarding/modern/ServerLoginPacketListenerBridge.java
@@ -4,6 +4,7 @@ import com.mojang.authlib.GameProfile;
 
 import dev.neuralnexus.taterapi.mc.world.entity.player.ProfilePublicKey;
 
+import org.adde0109.pcf.forwarding.ConnectionBridge;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
