+++ legacy/v12_2/src/forge/resources/pcf.mixins.v12_2.forge.json
@@ -8,7 +8,7 @@
   "plugin": "org.adde0109.pcf.mixin.plugin.PCFMixinPlugin",
   "package": "org.adde0109.pcf.mixin",
   "server": [
-    "v12_2.forge.forwarding.modern.ConnectionMixin",
+    "v12_2.forge.forwarding.ConnectionMixin",
     "v12_2.forge.forwarding.modern.ServerLoginPacketListenerImplHelloMixin",
     "v12_2.forge.forwarding.modern.ServerLoginPacketListenerImplMixin",
     "v12_2.forge.forwarding.modern.ServerLoginPacketListenerImplMixin$SLPLIMixin_12"
