+++ modern/v19_2/src/forge/java/org/adde0109/pcf/mixin/v19_2/forge/forwarding/ConnectionMixin.java
@@ -1,4 +1,4 @@
-package org.adde0109.pcf.mixin.v19_2.forge.forwarding.modern;
+package org.adde0109.pcf.mixin.v19_2.forge.forwarding;
 
 import dev.neuralnexus.taterapi.meta.Mappings;
 import dev.neuralnexus.taterapi.meta.anno.AConstraint;
@@ -12,7 +12,7 @@ import net.minecraft.network.Connection;
 import net.minecraft.network.ConnectionProtocol;
 import net.minecraft.network.PacketListener;
 
-import org.adde0109.pcf.forwarding.modern.ConnectionBridge;
+import org.adde0109.pcf.forwarding.ConnectionBridge;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
