+++ modern/v20_4/src/forge/resources/pcf.mixins.v20_4.forge.json
@@ -10,7 +10,7 @@
   "server": [
     "v20_4.forge.crossstitch.ArgumentNodeStubMixin",
     "v20_4.forge.crossstitch.ArgumentTypeInfoMixin",
-    "v20_4.forge.forwarding.modern.ConnectionMixin",
+    "v20_4.forge.forwarding.ConnectionMixin",
     "v20_4.forge.forwarding.modern.ServerLoginPacketListenerImplLoggerMixin",
     "v20_4.forge.forwarding.modern.ServerLoginPacketListenerImplMixin"
   ]
